requirements.txt
plotly==5.3.1
streamlit==1.3.0
requests==2.26.0
python-dotenv==0.19.1
